program-real-time-data-tracing/requirements.txt
PyQt5
matplotlib
pyserial
